+++ Cargo.toml
@@ -21,10 +21,10 @@ lto = true
 strip = true
 
 [profile.bench]
-# inherits = "release"
-debug = 1
-lto = false
-strip = false
+inherits = "release"
+# debug = 1
+# lto = false
+# strip = false
 
 [dev-dependencies]
 criterion = { version = "0.8", features = ["html_reports"] }
